+++ arch/arm/plat-omap/include/dspbridge/wmd.h
@@ -409,7 +409,7 @@ typedef void (*fxn_deh_notify) (struct deh_mgr *hdeh_mgr,
 typedef dsp_status(*fxn_chnl_open) (OUT struct chnl_object
 				    **phChnl,
 				    struct chnl_mgr *hchnl_mgr,
-				    short int chnl_mode,
+				    s8 chnl_mode,
 				    u32 uChnlId,
 				    CONST IN OPTIONAL struct
 				    chnl_attr * pattrs);
